Reorganize store into Rails-like folder structure
