+++ src/store/reducers/auth.test.js
@@ -3,9 +3,9 @@ import {
   REQUEST_LOGIN_SUCCESS,
   REQUEST_LOGIN_FAILURE,
   LOGOUT
-} from "./actions";
+} from "../actions/auth";
 
-import reducer from ".";
+import reducer from "./auth";
 
 let state;
